Rename ContentProviderStore into ContentProviderStoreCoreBase and move to cores package
